requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# OpenAPI parsing
pyyaml==6.0.1
jsonschema==4.21.1

# LLM integration
openai==2.8.1

# HTTP client for API calls
httpx==0.26.0
aiohttp==3.9.1

# Storage
aiosqlite==0.19.0

# Template rendering
jinja2==3.1.3

# Utilities
python-dotenv==1.0.0
uuid6==2024.1.12

# Demo UI
streamlit==1.31.0